feat(프로그래머스/level2/거리두기확인하기): 멀쩡한사각형 계산로직 수정. success.

+++ src/main/java/com/study/javaalgorithm/programmers/level2/멀쩡한사각형/Solution.java
@@ -8,20 +8,19 @@ public class Solution {
 		long gcd = calculateGreatestCommonFactor(max, min);
 
 		System.out.println("gcd = " + gcd);
-		long minW = (long)w / (long)gcd;
-		long minH = (long)h / (long)gcd;
-
+		long minW = (long)w / gcd;
+		long minH = (long)h / gcd;
 
 		long sliceSquareCount = minW + minH - 1;
 
 
-		return (long)w * (long)h - (long)(sliceSquareCount * gcd);
+		return (long)w * (long)h - (sliceSquareCount * gcd);
 	}
 
 	public long calculateGreatestCommonFactor(long w, long h) {
-		if((long)w % (long)h == 0) {
+		if(w % h == 0) {
 			return h;
 		}
-		return calculateGreatestCommonFactor(w, (long)w % (long)h);
+		return calculateGreatestCommonFactor(w, w % h);
 	}
 }
